update to_csv and move data models to separate module
